fix(agent): close RFC-64 discovery review boundaries
